Move tina folder (for some reason), and block mutations
