+++ components/blocks/content.tsx
@@ -3,7 +3,7 @@ import { Container } from "../util/container";
 import { Section } from "../util/section";
 import { TinaMarkdown } from "tinacms/dist/rich-text";
 import type { TinaTemplate } from "tinacms";
-import { PageBlocksContent } from "../../tina/__generated__/types";
+import { PageBlocksContent } from "../../.tina/__generated__/types";
 import { tinaField } from "tinacms/dist/react";
 
 export const Content = ({ data }: { data: PageBlocksContent }) => {
