+++ components/blocks/testimonial.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { Container } from "../util/container";
 import { Section } from "../util/section";
 import type { TinaTemplate } from "tinacms";
-import { PageBlocksTestimonial } from "../../tina/__generated__/types";
+import { PageBlocksTestimonial } from "../../.tina/__generated__/types";
 import { tinaField } from "tinacms/dist/react";
 
 export const Testimonial = ({ data }: { data: PageBlocksTestimonial }) => {
