+++ components/layout/layout.tsx
@@ -4,7 +4,7 @@ import { Header } from "./header";
 import { Footer } from "./footer";
 import { Theme } from "./theme";
 import layoutData from "../../content/global/index.json";
-import { Global } from "../../tina/__generated__/types";
+import { Global } from "../../.tina/__generated__/types";
 
 export const Layout = ({
   rawData = {},
