+++ package.json
@@ -27,7 +27,9 @@
   "dependencies": {
     "@headlessui/react": "1.7.5",
     "@tailwindcss/typography": "^0.5.8",
+    "@tinacms/datalayer": "^1.2.11",
     "date-fns": "^2.23.0",
+    "graphql": "15.8.0",
     "mongodb-level": "^0.0.2",
     "next": "^13.3.1",
     "react": "^18.2.0",
